requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
decouple==0.0.7
Django==4.1.7
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
whitenoise==6.3.0
wincertstore==0.2
